+++ disruptive/authentication.py
@@ -125,14 +125,14 @@ class Auth():
                 # This typically happens is credentials are fetched from
                 # the environment with a fallback to an empty string.
                 if len(credentials[key]) == 0:
-                    raise dterrors.ValueError(
+                    raise dterrors.EmptyStringError(
                         'Authentication credential <{}> is'
                         ' empty string.'.format(key)
                     )
 
             # If not, raise TypeError.
             else:
-                raise TypeError(
+                raise dterrors.TypeError(
                     'Authentication credential <{}> got type <{}>. '
                     'Expected <str>.'.format(
                         key, type(credentials[key]).__name__