+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes, fixes, and additions to the project is listed in this changelog.  
 After major version v1.0.0, the project adheres to [semantic versioning](https://semver.org/).
 
+# Unreleased
+### Fixed
+- [#49](https://github.com/disruptive-technologies/disruptive-python/pull/49) Fixed forgotten Humidity attribute rename from `temperature` to `celsius`.
+- [#50](https://github.com/disruptive-technologies/disruptive-python/pull/50) ConnectionStatus event parameter `available` should be `list[str]`, not `str`.
+
 # v0.4.0
 _Released on 2020-05-02._
 ### Changed